namada_state
Module ics23_specs

A module that contains

Functions

base_leaf_spec

Get the leaf spec for the base tree. The key is stored after hashing, but the stored value is the subtree’s root without hashing.

ibc_leaf_spec

Get the leaf spec for the ibc subtree.

ibc_proof_specs

Get the proof specs for ibc

leaf_spec

Get the leaf spec for the subtree. Non-hashed values are used for the verification with this spec because a subtree stores the key-value pairs after hashing.

proof_specs

Get the proof specs