namada_state
Trait DB

pub trait DB: Debug {
    type Cache;
    type WriteBatch: DBWriteBatch;
    type Migrator: DBUpdateVisitor<DB = Self>;
    type RestoreSource<'a>;

    // Required methods
    fn open(db_path: impl AsRef<Path>, cache: Option<&Self::Cache>) -> Self;
    fn restore_from(
        &mut self,
        source: Self::RestoreSource<'_>,
    ) -> Result<(), Error>;
    fn flush(&self, wait: bool) -> Result<(), Error>;
    fn read_last_block(&self) -> Result<Option<BlockStateRead>, Error>;
    fn add_block_to_batch(
        &self,
        state: BlockStateWrite<'_>,
        batch: &mut Self::WriteBatch,
        is_full_commit: bool,
    ) -> Result<(), Error>;
    fn read_block_header(
        &self,
        height: BlockHeight,
    ) -> Result<Option<BlockHeader>, Error>;
    fn read_merkle_tree_stores(
        &self,
        epoch: Epoch,
        base_height: BlockHeight,
        store_type: Option<StoreType>,
    ) -> Result<Option<MerkleTreeStoresRead>, Error>;
    fn has_replay_protection_entry(&self, hash: &Hash) -> Result<bool, Error>;
    fn read_subspace_val(&self, key: &Key) -> Result<Option<Vec<u8>>, Error>;
    fn read_subspace_val_with_height(
        &self,
        key: &Key,
        height: BlockHeight,
        last_height: BlockHeight,
    ) -> Result<Option<Vec<u8>>, Error>;
    fn read_diffs_val(
        &self,
        key: &Key,
        height: BlockHeight,
        is_old: bool,
    ) -> Result<Option<Vec<u8>>, Error>;
    fn write_subspace_val(
        &mut self,
        height: BlockHeight,
        key: &Key,
        value: impl AsRef<[u8]>,
        persist_diffs: bool,
    ) -> Result<i64, Error>;
    fn delete_subspace_val(
        &mut self,
        height: BlockHeight,
        key: &Key,
        persist_diffs: bool,
    ) -> Result<i64, Error>;
    fn batch() -> Self::WriteBatch;
    fn exec_batch(&self, batch: Self::WriteBatch) -> Result<(), Error>;
    fn batch_write_subspace_val(
        &self,
        batch: &mut Self::WriteBatch,
        height: BlockHeight,
        key: &Key,
        value: impl AsRef<[u8]>,
        persist_diffs: bool,
    ) -> Result<i64, Error>;
    fn batch_delete_subspace_val(
        &self,
        batch: &mut Self::WriteBatch,
        height: BlockHeight,
        key: &Key,
        persist_diffs: bool,
    ) -> Result<i64, Error>;
    fn prune_merkle_tree_store(
        &mut self,
        batch: &mut Self::WriteBatch,
        store_type: &StoreType,
        pruned_target: Either<BlockHeight, Epoch>,
    ) -> Result<(), Error>;
    fn read_bridge_pool_signed_nonce(
        &self,
        height: BlockHeight,
        last_height: BlockHeight,
    ) -> Result<Option<Uint>, Error>;
    fn write_replay_protection_entry(
        &mut self,
        batch: &mut Self::WriteBatch,
        key: &Key,
    ) -> Result<(), Error>;
    fn move_current_replay_protection_entries(
        &mut self,
        batch: &mut Self::WriteBatch,
    ) -> Result<(), Error>;
    fn prune_non_persisted_diffs(
        &mut self,
        batch: &mut Self::WriteBatch,
        height: BlockHeight,
    ) -> Result<(), Error>;
    fn overwrite_entry(
        &self,
        batch: &mut Self::WriteBatch,
        cf: &DbColFam,
        key: &Key,
        new_value: impl AsRef<[u8]>,
        persist_diffs: bool,
    ) -> Result<(), Error>;
    fn migrator() -> Self::Migrator;
    fn update_last_block_merkle_tree(
        &self,
        merkle_tree_stores: MerkleTreeStoresWrite<'_>,
        is_full_commit: bool,
    ) -> Result<(), Error>;

    // Provided method
    fn path(&self) -> Option<&Path> { ... }

}

A database backend.

Required Associated Types

type Cache

A DB’s cache

type WriteBatch: DBWriteBatch

A handle for batch writes

type Migrator: DBUpdateVisitor<DB = Self>

A type placeholder for DB migration implementation.

type RestoreSource<'a>

Source data to restore a database.

Required Methods

fn open(db_path: impl AsRef<Path>, cache: Option<&Self::Cache>) -> Self

Open the database from provided path

fn restore_from(&mut self, source: Self::RestoreSource<'_>) -> Result<(), Error>

Overwrite the contents of the current database with the data read from source.

fn flush(&self, wait: bool) -> Result<(), Error>

Flush data on the memory to persistent them

fn read_last_block(&self) -> Result<Option<BlockStateRead>, Error>

Read the last committed block’s metadata

fn add_block_to_batch( &self, state: BlockStateWrite<'_>, batch: &mut Self::WriteBatch, is_full_commit: bool, ) -> Result<(), Error>

Write block’s metadata. Merkle tree sub-stores are committed only when is_full_commit is true (typically on a beginning of a new epoch).

fn read_block_header( &self, height: BlockHeight, ) -> Result<Option<BlockHeader>, Error>

Read the block header with the given height from the DB

fn read_merkle_tree_stores( &self, epoch: Epoch, base_height: BlockHeight, store_type: Option<StoreType>, ) -> Result<Option<MerkleTreeStoresRead>, Error>

Read the merkle tree stores with the given epoch. If a store_type is given, it reads only the specified tree. Otherwise, it reads all trees, but some subtrees could be empty if the stores aren’t saved.

fn has_replay_protection_entry(&self, hash: &Hash) -> Result<bool, Error>

Check if the given replay protection entry exists

fn read_subspace_val(&self, key: &Key) -> Result<Option<Vec<u8>>, Error>

Read the latest value for account subspace key from the DB

fn read_subspace_val_with_height( &self, key: &Key, height: BlockHeight, last_height: BlockHeight, ) -> Result<Option<Vec<u8>>, Error>

Read the value for account subspace key at the given height from the DB. In our PersistentStorage (rocksdb), to find a value from arbitrary height requires looking for diffs from the given height, possibly up to the last_height.

fn read_diffs_val( &self, key: &Key, height: BlockHeight, is_old: bool, ) -> Result<Option<Vec<u8>>, Error>

Read the value for the account diffs at the corresponding height from the DB

fn write_subspace_val( &mut self, height: BlockHeight, key: &Key, value: impl AsRef<[u8]>, persist_diffs: bool, ) -> Result<i64, Error>

Write the value with the given height and account subspace key to the DB. Returns the size difference from previous value, if any, or the size of the value otherwise.

fn delete_subspace_val( &mut self, height: BlockHeight, key: &Key, persist_diffs: bool, ) -> Result<i64, Error>

Delete the value with the given height and account subspace key from the DB. Returns the size of the removed value, if any, 0 if no previous value was found.

fn batch() -> Self::WriteBatch

Start write batch.

fn exec_batch(&self, batch: Self::WriteBatch) -> Result<(), Error>

Execute write batch.

fn batch_write_subspace_val( &self, batch: &mut Self::WriteBatch, height: BlockHeight, key: &Key, value: impl AsRef<[u8]>, persist_diffs: bool, ) -> Result<i64, Error>

Batch write the value with the given height and account subspace key to the DB. Returns the size difference from previous value, if any, or the size of the value otherwise.

fn batch_delete_subspace_val( &self, batch: &mut Self::WriteBatch, height: BlockHeight, key: &Key, persist_diffs: bool, ) -> Result<i64, Error>

Batch delete the value with the given height and account subspace key from the DB. Returns the size of the removed value, if any, 0 if no previous value was found.

fn prune_merkle_tree_store( &mut self, batch: &mut Self::WriteBatch, store_type: &StoreType, pruned_target: Either<BlockHeight, Epoch>, ) -> Result<(), Error>

Prune Merkle tree stores at the given epoch

fn read_bridge_pool_signed_nonce( &self, height: BlockHeight, last_height: BlockHeight, ) -> Result<Option<Uint>, Error>

Read the signed nonce of Bridge Pool

fn write_replay_protection_entry( &mut self, batch: &mut Self::WriteBatch, key: &Key, ) -> Result<(), Error>

Write a replay protection entry

fn move_current_replay_protection_entries( &mut self, batch: &mut Self::WriteBatch, ) -> Result<(), Error>

Move the current replay protection bucket to the general one

fn prune_non_persisted_diffs( &mut self, batch: &mut Self::WriteBatch, height: BlockHeight, ) -> Result<(), Error>

Prune non-persisted diffs that are only kept for one block for rollback

fn overwrite_entry( &self, batch: &mut Self::WriteBatch, cf: &DbColFam, key: &Key, new_value: impl AsRef<[u8]>, persist_diffs: bool, ) -> Result<(), Error>

Overwrite a new value in storage, taking into account values stored at a previous height

fn migrator() -> Self::Migrator

Get an instance of DB migrator

fn update_last_block_merkle_tree( &self, merkle_tree_stores: MerkleTreeStoresWrite<'_>, is_full_commit: bool, ) -> Result<(), Error>

Update the merkle tree written for the committed last block

Provided Methods

fn path(&self) -> Option<&Path>

Get the path to the db in the filesystem, if it exists (the DB may be in-memory only)

Object Safety

This trait is not object safe.