namada_state
Trait DBIter

pub trait DBIter<'iter> {
    type PrefixIter: Debug + Iterator<Item = (String, Vec<u8>, Gas)>;
    type PatternIter: Debug + Iterator<Item = (String, Vec<u8>, Gas)>;

    // Required methods
    fn iter_prefix(&'iter self, prefix: Option<&Key>) -> Self::PrefixIter;
    fn iter_pattern(
        &'iter self,
        prefix: Option<&Key>,
        pattern: Regex,
    ) -> Self::PatternIter;
    fn iter_results(&'iter self) -> Self::PrefixIter;
    fn iter_old_diffs(
        &'iter self,
        height: BlockHeight,
        prefix: Option<&'iter Key>,
    ) -> Self::PrefixIter;
    fn iter_new_diffs(
        &'iter self,
        height: BlockHeight,
        prefix: Option<&'iter Key>,
    ) -> Self::PrefixIter;
    fn iter_current_replay_protection(&'iter self) -> Self::PrefixIter;
}

A database prefix iterator.

Required Associated Types

type PrefixIter: Debug + Iterator<Item = (String, Vec<u8>, Gas)>

Prefix iterator

type PatternIter: Debug + Iterator<Item = (String, Vec<u8>, Gas)>

Pattern iterator

Required Methods

fn iter_prefix(&'iter self, prefix: Option<&Key>) -> Self::PrefixIter

WARNING: This only works for values that have been committed to DB. To be able to see values written or deleted, but not yet committed, use the StorageWithWriteLog.

Read account subspace key value pairs with the given prefix from the DB, ordered by the storage keys.

fn iter_pattern( &'iter self, prefix: Option<&Key>, pattern: Regex, ) -> Self::PatternIter

WARNING: This only works for values that have been committed to DB. To be able to see values written or deleted, but not yet committed, use the StorageWithWriteLog.

Read account subspace key value pairs with the given pattern from the DB, ordered by the storage keys.

fn iter_results(&'iter self) -> Self::PrefixIter

Read results subspace key value pairs from the DB

fn iter_old_diffs( &'iter self, height: BlockHeight, prefix: Option<&'iter Key>, ) -> Self::PrefixIter

Read subspace old diffs at a given height

fn iter_new_diffs( &'iter self, height: BlockHeight, prefix: Option<&'iter Key>, ) -> Self::PrefixIter

Read subspace new diffs at a given height

fn iter_current_replay_protection(&'iter self) -> Self::PrefixIter

Read replay protection storage from the current bucket