namada_state
Trait StateRead

pub trait StateRead: StorageRead + Debug {
    type D: 'static + DB + for<'iter> DBIter<'iter>;
    type H: 'static + StorageHasher;

    // Required methods
    fn write_log(&self) -> &WriteLog;
    fn db(&self) -> &Self::D;
    fn in_mem(&self) -> &InMemory<Self::H>;
    fn charge_gas(&self, gas: Gas) -> Result<()>;

    // Provided methods
    fn db_has_key(&self, key: &Key) -> Result<(bool, Gas)> { ... }
    fn db_read(&self, key: &Key) -> Result<(Option<Vec<u8>>, Gas)> { ... }
    fn db_iter_prefix(
        &self,
        prefix: &Key,
    ) -> Result<(<Self::D as DBIter<'_>>::PrefixIter, Gas)> { ... }
    fn db_iter_results(&self) -> (<Self::D as DBIter<'_>>::PrefixIter, Gas) { ... }
    fn validity_predicate<Params: Keys>(
        &self,
        addr: &Address,
    ) -> Result<(Option<Hash>, Gas)> { ... }
    fn get_block_header(
        &self,
        height: Option<BlockHeight>,
    ) -> Result<(Option<BlockHeader>, Gas)> { ... }
}

Common trait for read-only access to write log, DB and in-memory state.

Required Associated Types

type D: 'static + DB + for<'iter> DBIter<'iter>

DB type

type H: 'static + StorageHasher

DB hasher type

Required Methods

fn write_log(&self) -> &WriteLog

Borrow WriteLog

fn db(&self) -> &Self::D

Borrow DB

fn in_mem(&self) -> &InMemory<Self::H>

Borrow InMemory state

fn charge_gas(&self, gas: Gas) -> Result<()>

Try to charge a given gas amount. Returns an error on out-of-gas.

Provided Methods

fn db_has_key(&self, key: &Key) -> Result<(bool, Gas)>

Check if the given key is present in storage. Returns the result and the gas cost.

fn db_read(&self, key: &Key) -> Result<(Option<Vec<u8>>, Gas)>

Returns a value from the specified subspace and the gas cost

fn db_iter_prefix( &self, prefix: &Key, ) -> Result<(<Self::D as DBIter<'_>>::PrefixIter, Gas)>

WARNING: This only works for values that have been committed to DB. To be able to see values written or deleted, but not yet committed, use the StorageWithWriteLog.

Returns a prefix iterator, ordered by storage keys, and the gas cost.

fn db_iter_results(&self) -> (<Self::D as DBIter<'_>>::PrefixIter, Gas)

Returns an iterator over the block results

fn validity_predicate<Params: Keys>( &self, addr: &Address, ) -> Result<(Option<Hash>, Gas)>

Get the hash of a validity predicate for the given account address and the gas cost for reading it.

fn get_block_header( &self, height: Option<BlockHeight>, ) -> Result<(Option<BlockHeader>, Gas)>

Get the block header

Object Safety

This trait is not object safe.