namada_state
Trait StorageRead

pub trait StorageRead {
    type PrefixIter<'iter>
       where Self: 'iter;

    // Required methods
    fn read_bytes(&self, key: &Key) -> Result<Option<Vec<u8>>, Error>;
    fn has_key(&self, key: &Key) -> Result<bool, Error>;
    fn iter_prefix<'iter>(
        &'iter self,
        prefix: &Key,
    ) -> Result<Self::PrefixIter<'iter>, Error>;
    fn iter_next<'iter>(
        &'iter self,
        iter: &mut Self::PrefixIter<'iter>,
    ) -> Result<Option<(String, Vec<u8>)>, Error>;
    fn get_chain_id(&self) -> Result<ChainId, Error>;
    fn get_block_height(&self) -> Result<BlockHeight, Error>;
    fn get_block_header(
        &self,
        height: BlockHeight,
    ) -> Result<Option<BlockHeader>, Error>;
    fn get_block_epoch(&self) -> Result<Epoch, Error>;
    fn get_pred_epochs(&self) -> Result<Epochs, Error>;
    fn get_tx_index(&self) -> Result<TxIndex, Error>;
    fn get_native_token(&self) -> Result<Address, Error>;

    // Provided methods
    fn read<T>(&self, key: &Key) -> Result<Option<T>, Error>
       where T: BorshDeserialize { ... }
    fn get_current_epoch_start_height(&self) -> Result<BlockHeight, Error> { ... }
    fn get_epoch_start_height(
        &self,
        epoch: Epoch,
    ) -> Result<Option<BlockHeight>, Error> { ... }
    fn get_epoch_at_height(
        &self,
        height: BlockHeight,
    ) -> Result<Option<Epoch>, Error> { ... }

}

Common storage read interface

Required Associated Types

type PrefixIter<'iter> where Self: 'iter

Storage read prefix iterator

Required Methods

fn read_bytes(&self, key: &Key) -> Result<Option<Vec<u8>>, Error>

Storage read raw bytes. It will try to read from the storage.

fn has_key(&self, key: &Key) -> Result<bool, Error>

Storage has_key in. It will try to read from the storage.

fn iter_prefix<'iter>( &'iter self, prefix: &Key, ) -> Result<Self::PrefixIter<'iter>, Error>

Storage prefix iterator ordered by the storage keys. It will try to get an iterator from the storage.

For a more user-friendly iterator API, use iter_prefix or iter_prefix_bytes instead.

fn iter_next<'iter>( &'iter self, iter: &mut Self::PrefixIter<'iter>, ) -> Result<Option<(String, Vec<u8>)>, Error>

Storage prefix iterator. It will try to read from the storage.

fn get_chain_id(&self) -> Result<ChainId, Error>

Getting the chain ID.

fn get_block_height(&self) -> Result<BlockHeight, Error>

Getting the block height. The height is that of the block to which the current transaction is being applied.

fn get_block_header( &self, height: BlockHeight, ) -> Result<Option<BlockHeader>, Error>

Getting the block header.

fn get_block_epoch(&self) -> Result<Epoch, Error>

Getting the block epoch. The epoch is that of the block to which the current transaction is being applied.

fn get_pred_epochs(&self) -> Result<Epochs, Error>

Given the information about predecessor block epochs

fn get_tx_index(&self) -> Result<TxIndex, Error>

Get the transaction index.

fn get_native_token(&self) -> Result<Address, Error>

Get the native token address

Provided Methods

fn read<T>(&self, key: &Key) -> Result<Option<T>, Error>
where T: BorshDeserialize,

Storage read Borsh encoded value. It will try to read from the storage and decode it if found.

fn get_current_epoch_start_height(&self) -> Result<BlockHeight, Error>

Get the height of the first block of the current epoch.

fn get_epoch_start_height( &self, epoch: Epoch, ) -> Result<Option<BlockHeight>, Error>

Get the height of the first block of the given epoch.

fn get_epoch_at_height( &self, height: BlockHeight, ) -> Result<Option<Epoch>, Error>

Given the epoch at the given block height.

Object Safety

This trait is not object safe.